matrices: add sympy.matrices.kind module
