+++ doc/src/modules/matrices/index.rst
@@ -13,6 +13,7 @@ Contents:
 
    matrices.rst
    matrixbase.rst
+   kind.rst
    dense.rst
    sparse.rst
    sparsetools.rst
