+++ doc/src/modules/matrices/kind.rst
@@ -1,0 +1,9 @@
+Matrix Kind
+===========
+
+.. module:: sympy.matrices.kind
+
+.. autoclass:: MatrixKind
+   :members:
+   :special-members:
+   :inherited-members:
