+++ sympy/matrices/expressions/matexpr.py
@@ -12,7 +12,8 @@ from sympy.external.gmpy import SYMPY_INTS
 from sympy.functions import conjugate, adjoint
 from sympy.functions.special.tensor_functions import KroneckerDelta
 from sympy.matrices.exceptions import NonSquareMatrixError
-from sympy.matrices.matrices import MatrixKind, MatrixBase
+from sympy.matrices.kind import MatrixKind
+from sympy.matrices.matrices import MatrixBase
 from sympy.multipledispatch import dispatch
 from sympy.utilities.misc import filldedent
