+++ sympy/matrices/expressions/tests/test_sets.py
@@ -6,7 +6,7 @@ from sympy.matrices.expressions.sets import MatrixSet
 from sympy.matrices.expressions.special import ZeroMatrix
 from sympy.testing.pytest import raises
 from sympy.sets.sets import SetKind
-from sympy.matrices.common import MatrixKind
+from sympy.matrices.kind import MatrixKind
 from sympy.core.kind import NumberKind
